Add SwiftLint and SwiftFormat but with no settings
The purpose of this change is to introduce the tools and show that they
have no effect with all rules disabled. The rules that are chosen can be
very contentious in any developer community, so this PR deliberately
avoids having any turned on and instead just enables the tools so that
rules can be handled in later PRs.