Perform changes synchronously (#940)
This is more of a guess to fix a crash, but it's not guaranteed, that it fixes the bug. When posting a new status, the app crashed occasionally when `try save()` in NSManagedObjectContext.swift. It seemed that the instance of NSManagedObjectContext just ... disappeared? The error-message was: EXC_BAD_INSTRUCTION (code=EXC_I386_INVOP, subcode=0x0)

So, instead of async black magic fuckery, just ... wait until it's done?