Never fetch complete followers and following list
Suspicion that this fetch was causing rate limit errors to appear almost immediately on launch for accounts with high numbers of followers/following. Since the actual list was never used, replace this with a notification that the list has changed, enabling the timeline to reload when a new follower is added or account blocked, etc.

Contributes to #1449