Better load experience
Prepare all rows before attempting to display anything, so that row heights don’t change dramatically after display.

Attempt to preserve the scroll position across navigation and relaunches. This is a work in progress, as it seems very difficult to acheive with SwiftUI. Here, we resort to splitting the known records into two parts, with the last read item at the beginning of the second chunk, which is the only chunk that gets displayed. Programmatically scrolling lands in the wrong place at least half the time, which is why we do not attempt to load all the waiting items at once unless the plan is to scroll to the top. Unfortunately, that makes for a very irritating experience when you want to read by scrolling slowly up, so this will need additional work.

Contributes to IOS-409
Contributes to IOS-10