Load only a slice of the available timeline into the VStack
VStack with a reasonable number of rows scrolls much smoother than LazyVStack and UITableView, but has performance problems as the row count rises. So now we only load a slice of the data at a time.

Also, we have to track scroll-related visibility ourselves because when using VStack, onAppear is called for every row as soon as the view is loaded no matter how far off-screen the rows are.

Contributes to IOS-10
Contributes to IOS-409