Simplify reloading and last read position handling
We are no longer attempting to track the last read and keep that at the top of the feed. The only attempts now made are to make the bottom and top loading experiences make sense, so the last item you saw should still be on screen after the reload.

Move the responsibility for reloading at the bottom from the last post view to the progress indicator view, so that you have a chance to read the last post in peace and have an indication that reloading is going to happen before it happens.

There is still some work to be done here, because the top loading scroll is highly unreliable. It will probably work better to give up on keeping the scroll position there and just treat it the same as pull to refresh.

Also, the bottom loading indicator looks like something is happening as soon as it appears, but the load will only be triggered when the bottom edge of its row actually comes into view.

Contributes to IOS-453