Add a nested content conceal model for quote posts with their own content warnings and filter hits
The top-level post may have its own content warning and filter results, so these have to be handled separately.

The quoted post only displays UI to reveal content, not to hide it after it has been revealed, so when the top level model hides its content, it sets the nested model to hide as well.