Keep view models for posts and notifications across reloads
Having the identifier for the row stay stable across reloads helps the system perform efficiently and give us some amount of scrolling stability automatically.

But when the content of the row changes, the view will only update if the view model updates its contents. Replacing the view model with a completely new instance, as we were doing, severs the relationship of the view model to the view and therefore fails to update the view.

Holding on to view models across reloads and updating the ones that we already have, rather than replacing them with brand new instances, solves the problem.

Fixes IOS-465