Improve layout of pickers with verbose options
Without the .fixedSize modifier, wrapped text in the picker label pushed the text of the whole element out of alignment with the background shape. With the .fixedSize modifier, everything stays vertically centered, although it is still possible for the label text to run outside the bounds of the background shape.

Fixes IOS-514