Ensure proper updating of currently visible posts, including recently deleted posts
This should all become much simpler when we move to iOS18 and can use the built in .scrollPosition modifier to handle the scroll stability.  But for now, we need to keep the currentDisplaySlice separate from the full results, and that can end up out of sync so that the updating done by the feed loader doesn’t affect every currently displayed item. Thus, we have the viewModel update the items it is currently displaying, although this may result in duplicated effort.