Interactive bottom loading indicator
Because the system does not offer a built-in pull-to-refresh from the bottom, we have to construct something ourselves. A fully custom animation would be nice since there is no way to tell the SwiftUI ProgressView to display without animating (the way the built-in pull-to-refresh behaves), but this is a compromise for now: we reveal the basic ProgressView bit by bit interactively as you scroll, while also increasing its size. Once the load action has been triggered, the full ProgressView remains visible.

Contributes to IOS-455