Fix problems with typing and autocomplete in the composer
The original method of converting a quote post to a link had unintended consequences because it did not go through the proper steps for updating the text view’s contents.

The contents are now correctly updated while keeping the cursor point valid.

A previously unnoticed bug has also been fixed, so that now choosing an autocomplete for a hashtag, mention, or custom emoji while inserting text in the middle of a post moves the cursor to the end of the autocompleted text as expected, rather than leaving it in the middle of the autocompletion.

Contributes to IOS-609