Ensure all available localizations are recognized
It seems iOS is not necessarily able to recognize available languages from .xcstrings files in packages rather than the main app bundle, which is a problem for us because our strings are provided in the MastodonSDK package’s Localizable.xcstrings file.

The solution here is to add a dummy Localizable.xcstrings file to the main bundle and ensure that each supported language provides a “translation” for the single placeholder key in that file. The contents of that “translation” doesn’t matter at all, because it will never be used in the UI. It is very important, however, that any new language be added to the .xcstrings file using the string catalog editing UI, which will automatically update the .pbxproj file’s “knownRegions” section, rather than by directly editing the json.

Automatic extraction of strings has been turned off because it is not useful to us given our package structure and manually-maintained L10Lookup struct.