Add beta profile view in SwiftUI and enable LiquidGlass
The beta profile view is off by default because it is display-only right now, none of the navigation or buttons function, and there is no editing capability.

Some of the main TimelineListView code has been cleaned up and refactored to allow the necessary scrolling behavior differences for the embedded timelines in the profile view. The embedded scrolling for the profile view works, although not as smoothly as it did in the UIKit version. The scroll gesture triggers the transition from scrollable to non-scrollable or vice versa at the proper point, but a new scroll gesture has to be started to actually scroll the now-scrollable area.

Liquid Glass has been turned on and some elements of the UI that did not make that transition well have been adjusted.

Contributes to IOS-592
Contributes to IOS-487