Update avatar button for Liquid Glass
Use the intended API for configuring the tab by setting a simple image rather than manipulating the view hierarchy of the tab bar item.

The image is now just a cicular crop of the avatar, with no border and no up-down chevron to indicate account switching, although the account switcher still works.

Contributes to IOS-624