Avoid triggering post actions while the view is scrolling
This was a problem when the timeline view was not the root view of the navigation controller, because the interactive pop gesture added by the system causes changes to gesture recognition that end up making the buttons hold on to the possibility of recognizing a tap long after scrolling starts. Tracking the scroll phase allows us to avoid doing an action while scrolling.

Fixes #1471